Converted ParserStage and ParserState to top-level objects.
