+++ features/events/syslog/src/main/java/org/opennms/netmgt/syslogd/RadixTreeParser.java
@@ -1,3 +1,31 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2017-2017 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2017 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
 package org.opennms.netmgt.syslogd;
 
 import java.nio.ByteBuffer;
@@ -10,8 +38,6 @@ import org.opennms.core.collections.RadixTree;
 import org.opennms.core.collections.RadixTreeImpl;
 import org.opennms.core.collections.RadixTreeNode;
 import org.opennms.netmgt.model.events.EventBuilder;
-import org.opennms.netmgt.syslogd.ParserStageSequenceBuilder.ParserStage;
-import org.opennms.netmgt.syslogd.ParserStageSequenceBuilder.ParserState;
 import org.opennms.netmgt.xml.event.Event;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -60,7 +86,7 @@ public class RadixTreeParser implements ByteBufferParser<Event> {
 
 	@Override
 	public CompletableFuture<Event> parse(ByteBuffer incoming) {
-		ParserState state = new ParserState(incoming, new EventBuilder("uei.opennms.org/test", this.getClass().getSimpleName()));
+		ParserState state = new ParserState(incoming, new EventBuilder());
 
 		// TODO: Use better collection than ArrayList?
 		final List<CompletableFuture<ParserState>> finishedFutures = new ArrayList<>();
